[EI-226] [web] refactor: Refactor indicator-related hooks to use the new numerical-guidance folder structure
