+++ web/app/business/hooks/numerical-guidance/indicator-board/use-indicator-board.hook.ts
