+++ web/app/ui/components/numerical-guidance/indicator-board-metadata/metadata-list/metadata-list-item.tsx
@@ -11,7 +11,7 @@ import DraggableItem, { Item } from '../../../view/atom/draggable/draggable-item
 import { useIndicatorBoardMetadataViewModel } from '@/app/business/hooks/numerical-guidance/indicator-board-metedata/use-indicator-board-metadata-view-model.hook';
 import { useEffect, useState } from 'react';
 import { cn } from '@/app/utils/style';
-import { useIndicatorBoard } from '@/app/business/hooks/indicator-board/use-indicator-board.hook';
+import { useIndicatorBoard } from '@/app/business/hooks/numerical-guidance/indicator-board/use-indicator-board.hook';
 import IndicatorUnitSelector from './indicator-unit-selector';
 
 type MetadataListItemProps = {
