+++ web/app/ui/components/numerical-guidance/indicator/date-range-navigator.tsx
@@ -1,4 +1,4 @@
-import { useIndicatorBoard } from '@/app/business/hooks/indicator-board/use-indicator-board.hook';
+import { useIndicatorBoard } from '@/app/business/hooks/numerical-guidance/indicator-board/use-indicator-board.hook';
 import ToggleGroup from '../../view/molecule/toggle-group';
 import { DateRange, dateRange } from '@/app/store/stores/numerical-guidance/indicator-board.store';
